pyproject.toml
[tool.poetry]
name = "money_bot"
version = "0.1.0"
description = ""
authors = ["erastoff <yuriy.erastov@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
aiogram = "^3.1.1"
loguru = "^0.7.2"
pydantic-settings = "^2.0.3"
python-decouple = "^3.8"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
aioredis = "^2.0.1"
python-binance = "^1.0.19"

[tool.poetry.dev-dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
